refactor: allow multiple Action routes
