+++ apps/ui/src/helpers/auction/index.ts
@@ -2,14 +2,16 @@ import { ApolloClient, InMemoryCache } from '@apollo/client/core';
 import { AuctionDetailFragment } from './gql/graphql';
 import { auctionQuery } from './queries';
 
-const SUBGRAPH_URLS = {
+export type AuctionNetworkId = 'eth' | 'sep';
+
+const SUBGRAPH_URLS: Record<AuctionNetworkId, string> = {
   sep: 'https://subgrapher.snapshot.org/subgraph/arbitrum/Hs3FN65uB3kzSn1U5kPMrc1kHqaS9zQMM8BCVDwNf7Fn',
   eth: 'https://subgrapher.snapshot.org/subgraph/arbitrum/98f9T2v1KtNnZyexiEgNLMFnYkXdKoZq9Pt1EYQGq5aH'
-} as const;
+};
 
 export async function getAuction(
   id: string,
-  network: string
+  network: AuctionNetworkId
 ): Promise<{ auctionDetail: AuctionDetailFragment } | null> {
   const subgraphUrl = SUBGRAPH_URLS[network];
   if (!subgraphUrl) throw new Error(`Unknown network: ${network}`);
