+++ apps/ui/src/routes/default.ts
@@ -1,4 +1,6 @@
 import App from '@/views/App.vue';
+import AuctionBids from '@/views/Auction/Bids.vue';
+import AuctionOverview from '@/views/Auction/Overview.vue';
 import Auction from '@/views/Auction.vue';
 import CreateSpaceSnapshot from '@/views/CreateSpaceSnapshot.vue';
 import CreateSpaceSnapshotX from '@/views/CreateSpaceSnapshotX.vue';
@@ -58,7 +60,11 @@ export default [
   {
     path: '/auction/:id',
     name: 'auction',
-    component: Auction
+    component: Auction,
+    children: [
+      { path: '', name: 'auction-overview', component: AuctionOverview },
+      { path: 'bids', name: 'auction-bids', component: AuctionBids }
+    ]
   },
   {
     path: '/settings',
