+++ apps/ui/src/views/Auction/Bids.vue
@@ -1,0 +1,1 @@
+<template>Bids</template>
